% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{AFLP_filter}
\alias{AFLP_filter}
\title{Filters AFLP matrices to remove any loci that are absent from included samples}
\usage{
AFLP_filter(infile, outfile)
}
\arguments{
\item{\code{infile}}{filename of Structure dataset}

\item{\code{outfile}}{desired filename for filtered Structure dataset}
}
\description{
This function reads an \code{STRUCTURE} formatted AFLP dataset and writes a filtered version of that file
}
\examples{
AFLP_filter(myStructureFile,filteredStructureFile)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{filtering}

