% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{errorRate}
\alias{errorRate}
\title{Calculates mismatch error rate for AFLP dataset}
\usage{
errorRate(filename, duplicates)
}
\arguments{
\item{\code{filename}}{filename of matrix containing duplicates}

\item{\code{duplicates}}{integer, the number of duplicate pairs}
}
\description{
This function reads an \code{STRUCTURE} formatted AFLP dataset and writes a filtered version of that file
}
\examples{
errorRate(myAflpMatrix,duplicates=10)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{error}

