% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{filter.loci}
\alias{filter.loci}
\title{Filters AFLP matrices to remove any loci that are absent from included samples}
\usage{
filter.loci(infile, outfile)
}
\arguments{
\item{\code{infile}}{filename of Structure dataset}

\item{\code{outfile}}{desired filename for filtered Structure dataset}
}
\description{
This function reads a \code{STRUCTURE} formatted AFLP dataset and writes a filtered version of that file
}
\examples{
filter.loci(myStructureFile,filteredStructureFile)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{filtering}

