% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{filter.samples}
\alias{filter.samples}
\title{Filters AFLP matrices to remove poor quality samples}
\usage{
filter.samples(infile, outfile, cutoff = 1)
}
\arguments{
\item{\code{infile}}{filename of Structure dataset}

\item{\code{outfile}}{desired filename for filtered Structure dataset}

\item{\code{cutoff}}{number of standard deviations below the mean to use as a threshold}
}
\description{
This function reads a \code{STRUCTURE}-formatted AFLP dataset and writes a filtered version of that file
}
\examples{
filter.samples(myStructureFile,filteredStructureFile, cutoff)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{filtering}

