% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{makeTable}
\alias{makeTable}
\title{Create AFLP data matrix from raw peak heights}
\usage{
makeTable(filename, binthresh, allelethresh)
}
\arguments{
\item{\code{filename}}{filename of matrix containing duplicates}

\item{\code{duplicates}}{integer, the number of duplicate pairs}
}
\description{
This function creates AFLP data matrix from raw peak heights scored using
user determined locus and phenotype thresholds.
This  function borrows heavily from AFLPscore v1.4 by Raj Whitlock
Whitlock et al. 2008 Mol Ecol Res. 8:725-735
shef.ac.uk/molecol/software/aflpscore
}
\examples{
makeTable(myRawPeakHeights,binthresh=100,allelethresh=100)
}
\keyword{AFLP,}
\keyword{AFLPscore}

