% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{RawGeno_import}
\alias{RawGeno_import}
\title{This function imports the binary table output from RawGeno}
\usage{
RawGeno_import(outfilename)
}
\arguments{
\item{\code{outfilename}}{desired filename for peak height files}
}
\description{
This function converts the binary table generated by an active session of RawGeno
to a format that can be read by other functions in this package
}
\examples{
RawGeno_import(outfile.txt)
}
\keyword{AFLP,}
\keyword{RawGeno}
\keyword{STRUCTURE,}

