% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read.structure}
\alias{read.structure}
\title{Read a standard STRUCTURE dataset}
\usage{
read.structure(infile)
}
\arguments{
\item{\code{infile}}{filename of Structure dataset}
}
\description{
This function reads an \code{STRUCTURE} formatted AFLP dataset and returns a row labelled dataframe
}
\examples{
read.structure(myStructureFile)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{read}

