% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{write.structure}
\alias{write.structure}
\title{Write a standard fomatted STRUCTURE dataset}
\usage{
write.structure(dat, outfile)
}
\arguments{
\item{\code{dat}}{a dataframe containing AFLP data generated by read.structure}

\item{\code{outfile}}{desired filename for filtered Structure dataset}
}
\description{
This function reads a \code{STRUCTURE}-formatted AFLP dataset and returns a row labelled dataframe
}
\examples{
write.structure(dat, myOutfile)
}
\keyword{AFLP,}
\keyword{STRUCTURE,}
\keyword{write}

